feat: skeletons of running process of style transfer feat: ongoing on build VGG16_pt model and style_transfer model refactor: rename losses.py to loss.py chore: add __pycache__ on .gitignore
